PROJQUAY-7306: refactor(quota): fold QuotaNotificationWorker into QuotaTotalWorker
Eliminates a standalone supervisord process by moving periodic quota
notification scanning into QuotaTotalWorker as a second add_operation,
following the GC worker pattern for image expiry notifications.
