+++ conf/init/supervisord_conf_create.py
@@ -58,7 +58,6 @@ def registry_services():
         "securityscanningnotificationworker": {"autostart": "true"},
         "quotatotalworker": {"autostart": "true"},
         "quotaregistrysizeworker": {"autostart": "true"},
-        "quotanotificationworker": {"autostart": "true"},
         "autopruneworker": {"autostart": "true"},
         "proxycacheblobworker": {"autostart": "true"},
         "pullstatsredisflushworker": {"autostart": "true"},
@@ -100,7 +99,6 @@ def config_services():
         "securityscanningnotificationworker": {"autostart": "false"},
         "quotatotalworker": {"autostart": "false"},
         "quotaregistrysizeworker": {"autostart": "false"},
-        "quotanotificationworker": {"autostart": "false"},
         "autopruneworker": {"autostart": "false"},
         "proxycacheblobworker": {"autostart": "false"},
         "pullstatsredisflushworker": {"autostart": "false"},
